# openapi/v1.yaml
components:
  responses:
    Forbidden:
      content: &id001
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      description: Authenticated but not allowed (role, scope, tenant, quota).
    InternalError:
      content: *id001
      description: Unexpected error (a fixed body; details stay in the server log).
    Invalid:
      content: *id001
      description: The request is not valid.
    NotFound:
      content: *id001
      description: Not found (also for resources of another tenant).
    OtherError:
      content: *id001
      description: Any other error status, e.g. 409 conflict, 410 gone, 413 too large.
    TooManyRequests:
      content: *id001
      description: Rate limit or quota exceeded.
      headers:
        Retry-After:
          schema:
            minimum: 1
            type: integer
    Unauthorized:
      content: *id001
      description: Missing, invalid, expired or revoked credential.
      headers:
        WWW-Authenticate:
          schema:
            type: string
  schemas:
...
    CompleteIn:
      properties:
        sha256:
          description: client SHA-256 (hex)
          maxLength: 64
          minLength: 64
          title: Sha256
          type: string
      required:
      - sha256
      title: CompleteIn
      type: object
...
    DeviceIn:
      properties:
        name:
          maxLength: 100
          minLength: 1
          title: Name
          type: string
        public_key:
          description: Ed25519 public key, 32 bytes, base64
          title: Public Key
          type: string
      required:
      - name
      - public_key
      title: DeviceIn
      type: object
    DeviceOut:
      properties:
        alg:
          title: Alg
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        id:
          format: uuid
          title: Id
          type: string
        name:
          title: Name
          type: string
        public_key:
          title: Public Key
          type: string
        revoked_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Revoked At
      required:
      - id
      - name
      - alg
      - public_key
      - created_at
      - revoked_at
      title: DeviceOut
      type: object
...
    ReceivedPartOut:
      properties:
        part_number:
          title: Part Number
          type: integer
        sha256:
          title: Sha256
          type: string
        size_bytes:
          title: Size Bytes
          type: integer
      required:
      - part_number
      - size_bytes
      - sha256
      title: ReceivedPartOut
      type: object
...
    UploadIn:
      properties:
        filename:
          maxLength: 200
          minLength: 1
          title: Filename
          type: string
        part_size:
          anyOf:
          - exclusiveMinimum: 0.0
            type: integer
          - type: 'null'
          title: Part Size
        session_id:
          format: uuid
          title: Session Id
          type: string
        size_bytes:
          exclusiveMinimum: 0.0
          title: Size Bytes
          type: integer
        synthetic:
          default: false
          description: 'SEC-071: true for synthetic or licence-checked public data (required outside prod
            and for synthetic-only tenants)'
          title: Synthetic
          type: boolean
      required:
      - session_id
      - filename
      - size_bytes
      title: UploadIn
      type: object
    UploadOut:
      properties:
        completed_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Completed At
        created_at:
          format: date-time
          title: Created At
          type: string
        dataset_id:
          format: uuid
          title: Dataset Id
          type: string
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        filename:
          title: Filename
          type: string
        id:
          format: uuid
          title: Id
          type: string
        n_parts:
          title: N Parts
          type: integer
        part_size:
          title: Part Size
          type: integer
        parts:
          items:
            $ref: '#/components/schemas/PartTargetOut'
          title: Parts
          type: array
        received_parts:
          items:
            $ref: '#/components/schemas/ReceivedPartOut'
          title: Received Parts
          type: array
        recording_ids:
          items:
            type: string
          title: Recording Ids
          type: array
        server_sha256:
          anyOf:
          - type: string
          - type: 'null'
          title: Server Sha256
        session_id:
          format: uuid
          title: Session Id
          type: string
        size_bytes:
          title: Size Bytes
          type: integer
        state:
          title: State
          type: string
        synthetic:
          title: Synthetic
          type: boolean
      required:
      - id
      - dataset_id
      - session_id
      - filename
      - size_bytes
      - part_size
      - n_parts
      - state
      - synthetic
      - parts
      - received_parts
      - server_sha256
      - error
      - recording_ids
      - created_at
      - completed_at
      title: UploadOut
      type: object
...
  securitySchemes:
    apiKey:
      bearerFormat: nfb_live_<public id>_<secret>
      description: Scoped, expiring API key (SEC-014). Revocation is effective at once for new requests
        and within 60 s for open streams (SEC-017).
      scheme: bearer
      type: http
    oidc:
      description: 'User access token from the identity provider (Authorization: Bearer).'
      openIdConnectUrl: https://id.example.invalid/.well-known/openid-configuration
      type: openIdConnect
info:
  description: 'Platform API v1. Data flows device -> SDK -> platform only: no operation, event or field
    sends anything to acquisition hardware (SEC-090). Errors are RFC 9457 application/problem+json. Only
    additive changes within v1; deprecations carry Deprecation and Sunset headers.'
  title: Platform API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/datasets/{dataset_id}/uploads:
    post:
      operationId: createUpload
      parameters:
      - in: path
        name: dataset_id
        required: true
        schema:
          format: uuid
          title: Dataset Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadIn'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadOut'
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: Create Upload
      x-nf-action: upload:create
      x-nf-api-key-scope: data:write
...
  /v1/devices:
    get:
      operationId: listDevices
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 200
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DeviceOut'
                title: Response Listdevices
                type: array
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: List Devices
      x-nf-action: device:read
      x-nf-api-key-scope: data:write
    post:
      description: 'SEC-016: the key pair is generated on the device; only the public key is registered.'
      operationId: registerDevice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceIn'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceOut'
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: Register Device
      x-nf-action: device:create
      x-nf-api-key-scope: data:write
...
  /v1/recordings/{recording_id}/data:
    get:
      description: 'Read the window ``[start, end)`` seconds of a recording (step 2.4).


        Decrypted server-side with the subject''s DEK. Response format ``nf-window/1`` (see

        ``nf_platform.ingest.recording_store``): a header (channels, units, scale/offset, dtype,

        shape [channels, samples], level, decimation, sfreq, start_s, chunk_id) plus the samples,

        as JSON (``format=json``, at most ``window_json_max_values`` values) or as

        ``application/vnd.nf.window.v1`` bytes (``format=binary``, at most ``window_max_values``).

        Larger windows get 422 ``window-too-large``: use a pyramid ``level`` or a shorter window.

        Quarantined recordings: 403 except for owner/admin/data-steward.'
      operationId: readRecordingData
      parameters:
      - in: path
        name: recording_id
        required: true
        schema:
          format: uuid
          title: Recording Id
          type: string
      - description: window start [s]
        in: query
        name: start
        required: true
        schema:
          description: window start [s]
          minimum: 0
          title: Start
          type: number
      - description: window end [s], exclusive
        in: query
        name: end
        required: true
        schema:
          description: window end [s], exclusive
          exclusiveMinimum: 0
          title: End
          type: number
      - description: comma-separated names or indices
        in: query
        name: channels
        required: false
        schema:
          anyOf:
          - maxLength: 65536
            type: string
          - type: 'null'
          description: comma-separated names or indices
          title: Channels
      - in: query
        name: level
        required: false
        schema:
          default: 0
          maximum: 11
          minimum: 0
          title: Level
          type: integer
      - in: query
        name: kind
        required: false
        schema:
          default: mean
          enum:
          - mean
          - min
          - max
          title: Kind
          type: string
      - in: query
        name: format
        required: false
        schema:
          default: json
          enum:
          - json
          - binary
          title: Format
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WindowJsonOut'
            application/vnd.nf.window.v1:
              schema:
                format: binary
                type: string
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: Read Recording Data
      x-nf-action: signal:read
      x-nf-api-key-scope: data:read
...
  /v1/sessions/{session_id}/streams:
    post:
      operationId: openStream
      parameters:
      - in: path
        name: session_id
        required: true
        schema:
          format: uuid
          title: Session Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StreamIn'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamOut'
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: Open Stream
      x-nf-action: stream:create
      x-nf-api-key-scope: data:write
...
  /v1/uploads/{upload_id}:
    get:
      operationId: getUpload
      parameters:
      - in: path
        name: upload_id
        required: true
        schema:
          format: uuid
          title: Upload Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadOut'
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: Get Upload
      x-nf-action: upload:read
      x-nf-api-key-scope: data:write
  /v1/uploads/{upload_id}/complete:
    post:
      operationId: completeUpload
      parameters:
      - in: path
        name: upload_id
        required: true
        schema:
          format: uuid
          title: Upload Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteIn'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadOut'
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: Complete Upload
      x-nf-action: upload:create
      x-nf-api-key-scope: data:write
  /v1/uploads/{upload_id}/parts/{part_number}:
    put:
      description: 'Local signed-URL shim: the part body is sealed under the subject DEK before storage.'
      operationId: putUploadPart
      parameters:
      - in: path
        name: upload_id
        required: true
        schema:
          format: uuid
          title: Upload Id
          type: string
      - in: path
        name: part_number
        required: true
        schema:
          maximum: 10000
          minimum: 1
          title: Part Number
          type: integer
      requestBody:
        content:
          application/octet-stream: {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceivedPartOut'
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: Put Upload Part
      x-nf-action: upload:create
      x-nf-api-key-scope: data:write
...
servers:
- url: https://api.example.invalid
...